- Fix some MSVC 2005 Compilation warnings.

+++ sf-cvs/trunk/audacity-src/src/WaveTrack.cpp
@@ -1418,7 +1418,7 @@ int WaveTrack::GetClipIndex(WaveClip* clip)
 
 WaveClip* WaveTrack::GetClipByIndex(int index)
 {
-   if(index < mClips.GetCount())
+   if(index < (int)mClips.GetCount())
       return mClips.Item(index)->GetData();
    else
       return NULL;
